+++ CHANGELOG.md
@@ -9,13 +9,14 @@
 
 - **New Feature**: Reworked "Edit" interface for the summary prompt. This is now where the `{{history}}`, `{{message}}`, and `{{words}}` macro are defined and modified. You can also create custom macros for the summary prompt using either a message range or STScript.
 - **New Feature**: A new setting called `Static Memory Mode` can be enabled in the general injection settings. This mode makes long-term memories always be injected separately from short-term memories, regardless of context. Disabled by default.
+- **New Feature**: New setting next to `Summarization Time Delay` called `Skip First` determines whether the first summary right after a character message will be delayed or not when auto-summarizing (default false). Turning this on will be the same as the old behavior.
 - **New Slash Command**: `/qm-max-summary-tokens` returns the max tokens allowed for summarization given the current completion preset.
 - **New Slash Command**: `/qm-set` allows you to set the memory for a message.
 - **Changed**: All slash commands now use "-" instead of "_" and start with "qm-" (with "qvink-memory-" as an alias) to avoid overlapping with other commands. See the README for the updated list of all commands.
 - **Changed**: All CSS variables now use "-" instead of "_", and are prefixed with "--qm-"
 - **Changed**: Messages that have been removed from context are now turned grey in the chat. This can be modified by using the `--qm-message-removed` CSS variable.
 - **Changed**: Summaries that aren't included in memory can now be changed style with the `--qm-default` CSS variable.
-- **Changed**: The slash command `/qm-get` (previously `/get_memory`) can now accept a range of indexes and a custom separator
+- **Changed**: The slash command `/qm-get` (previously `/get_memory`) can now accept a range of indexes and a custom separator.
 - **Changed**: Modified how profile changes are detected under the hood. Shouldn't affect anything.
 - **Changed**: You won't believe it I moved settings around again. Auto-Summarization settings now get their own section.
 - **Fixed**: The config popout now works with MovingUI